Set ALBUMS ID as float. Add spaces after commas

+++ src/db.py
@@ -573,7 +573,7 @@ class DB:
                 ,COUNTRY text    \
                 ,COMMENT text    \
                 ,SEARCH  text    \
-                ,RATING  integer \
+                ,RATING  float   \
                                                    )'
         try:
             self.dbc.execute(query)
